+++ Packy/Data/SocialLogin/Controller/KakaoLoginController.swift
@@ -1,0 +1,8 @@
+//
+//  KakaoLoginController.swift
+//  Packy
+//
+//  Created by Mason Kim on 1/8/24.
+//
+
+import Foundation
